requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pymongo
python-dotenv
fastapi
uvicorn[standard]
pydantic
python-multipart
langchain-openai
PyJWT
passlib[bcrypt]
python-jose[cryptography]
